dcase2024_task1_baseline-main/requirements.txt
pytorch-lightning>=2.0.0
torchinfo
transformers
pandas
scikit-learn
wandb
av
h5py
sox
soundfile

hydra-core
omegaconf

pyyaml
